Adding Motor Controller Board Eagle Project
This adds the Eagle files for the Motor Controller Board.

L6742LIB.lbr holds several footprints and custom components needed by
the project.

L6472V1.brd is the board file for laying out the PCB.

L6472V1.sch is the schematic file for the board.

eagle.epf is a file created by Eagle. Not sure what it does but it seems
important...